+++ packages/arb-token-bridge-ui/package.json
@@ -10,7 +10,7 @@
     "@headlessui/react": "^1.7.8",
     "@headlessui/tailwindcss": "^0.1.2",
     "@heroicons/react": "^1.0.3",
-    "@rainbow-me/rainbowkit": "^0.12.9",
+    "@rainbow-me/rainbowkit": "0.12.13",
     "@rehooks/local-storage": "^2.4.4",
     "@sentry/react": "^7.45.0",
     "@sentry/tracing": "^7.45.0",
