+++ packages/arb-token-bridge-ui/src/components/TransactionHistory/TransactionsTable/TransactionsTable.tsx
@@ -75,7 +75,6 @@ export type TransactionsTableProps = {
   transactions: MergedTransaction[]
   loading: boolean
   error: boolean
-  pendingTransactions: MergedTransaction[]
 }
 
 export function TransactionsTable({
@@ -84,8 +83,7 @@ export function TransactionsTable({
   setPageParams,
   transactions,
   loading,
-  error,
-  pendingTransactions
+  error
 }: TransactionsTableProps) {
   const { isSmartContractWallet } = useNetworksAndSigners()
 
@@ -140,15 +138,15 @@ export function TransactionsTable({
     return [...newerTransactions.reverse(), ...subgraphTransactions]
   }, [transactions, localTransactionsKey])
 
-  const pendingTransactionsMap = useMemo(() => {
-    // map of all the locally-stored pending transactions
+  const locallyStoredTransactionsMap = useMemo(() => {
+    // map of all the locally-stored transactions (pending + recently executed)
     // so that tx rows can easily subscribe to live-local status without refetching table data
-    const pendingTxMap = new Map<string, MergedTransaction>()
-    pendingTransactions.forEach(tx => {
-      pendingTxMap.set(tx.txId, tx)
+    const localTxMap = new Map<string, MergedTransaction>()
+    locallyStoredTransactions.forEach(tx => {
+      localTxMap.set(tx.txId, tx)
     })
-    return pendingTxMap
-  }, [localTransactionsKey])
+    return localTxMap
+  }, [locallyStoredTransactions])
 
   const status = (() => {
     if (loading) return TableStatus.LOADING
@@ -175,7 +173,7 @@ export function TransactionsTable({
 
       {
         <table className="w-full overflow-hidden  rounded-b-lg bg-white">
-          <thead className="text-left text-sm text-gray-7">
+          <thead className="text-left text-sm text-gray-10">
             <tr>
               <th className="py-3 pl-6 pr-3 font-normal">Status</th>
               <th className="px-3 py-3 font-normal">Time</th>
@@ -216,18 +214,19 @@ export function TransactionsTable({
               _transactions.map((tx, index) => {
                 const isLastRow = index === _transactions.length - 1
 
-                // if transaction is present in pending transactions, subscribe to that in this row,
-                // this will make sure the row updates with any updates in the pending tx state
+                // if transaction is present in local (pending + recently executed) transactions, subscribe to that in this row,
+                // this will make sure the row updates with any updates in the local app state
                 // else show static subgraph table data
-                const pendingTransaction = pendingTransactionsMap.get(tx.txId)
-                const finalTx = pendingTransaction ? pendingTransaction : tx
+                const locallyStoredTransaction =
+                  locallyStoredTransactionsMap.get(tx.txId)
+                const finalTx = locallyStoredTransaction ?? tx
 
                 if (isDeposit(finalTx)) {
                   return (
                     <TransactionsTableDepositRow
                       key={`${finalTx.txId}-${finalTx.direction}`}
                       tx={finalTx}
-                      className={!isLastRow ? 'border-b border-gray-3' : ''}
+                      className={!isLastRow ? 'border-b border-gray-5' : ''}
                     />
                   )
                 } else if (isWithdrawal(finalTx)) {
@@ -235,7 +234,7 @@ export function TransactionsTable({
                     <TransactionsTableWithdrawalRow
                       key={`${finalTx.txId}-${finalTx.direction}`}
                       tx={finalTx}
-                      className={!isLastRow ? 'border-b border-gray-3' : ''}
+                      className={!isLastRow ? 'border-b border-gray-5' : ''}
                     />
                   )
                 } else {
