+++ packages/arb-token-bridge-ui/src/token-bridge-sdk/index.ts
@@ -16,17 +16,4 @@ export type {
   L2ToL1MessageData
 } from './hooks/useTransactions'
 
-export { ERC20__factory } from '@arbitrum/sdk/dist/lib/abi/factories/ERC20__factory'
-
-export {
-  getDefaultTokenName,
-  getDefaultTokenSymbol,
-  getL1TokenData,
-  getL1TokenAllowance,
-  getL2TokenData,
-  validateTokenList
-} from './util/index'
-export { getL1ERC20Address } from './util/getL1ERC20Address'
-export { getUniqueIdOrHashFromEvent } from './util/migration'
-
 export { useGasPrice } from './hooks/useGasPrice'
